+++ Modules/Microsoft365DSC/DSCResources/MSFT_EXOActiveSyncDeviceAccessRule/MSFT_EXOActiveSyncDeviceAccessRule.psm1
@@ -119,7 +119,7 @@ function Get-TargetResource
             }
             elseif ($null -ne $GlobalAdminAccount)
             {
-                $tenantIdValue = $GlobalAdminAccount.UserName.Split('@')[0]
+                $tenantIdValue = $GlobalAdminAccount.UserName.Split('@')[1]
             }
             Add-M365DSCEvent -Message $_ -EntryType 'Error' `
                 -EventID 1 -Source $($MyInvocation.MyCommand.Source) `
@@ -417,7 +417,7 @@ function Export-TargetResource
             }
             elseif ($null -ne $GlobalAdminAccount)
             {
-                $tenantIdValue = $GlobalAdminAccount.UserName.Split('@')[0]
+                $tenantIdValue = $GlobalAdminAccount.UserName.Split('@')[1]
             }
             Add-M365DSCEvent -Message $_ -EntryType 'Error' `
                 -EventID 1 -Source $($MyInvocation.MyCommand.Source) `