+++ Modules/Microsoft365DSC/DSCResources/MSFT_ODSettings/MSFT_ODSettings.psm1
@@ -199,7 +199,7 @@ function Get-TargetResource
             }
             elseif ($null -ne $GlobalAdminAccount)
             {
-                $tenantIdValue = $GlobalAdminAccount.UserName.Split('@')[0]
+                $tenantIdValue = $GlobalAdminAccount.UserName.Split('@')[1]
             }
             Add-M365DSCEvent -Message $_ -EntryType 'Error' `
                 -EventID 1 -Source $($MyInvocation.MyCommand.Source) `
@@ -610,7 +610,7 @@ function Export-TargetResource
             }
             elseif ($null -ne $GlobalAdminAccount)
             {
-                $tenantIdValue = $GlobalAdminAccount.UserName.Split('@')[0]
+                $tenantIdValue = $GlobalAdminAccount.UserName.Split('@')[1]
             }
             Add-M365DSCEvent -Message $_ -EntryType 'Error' `
                 -EventID 1 -Source $($MyInvocation.MyCommand.Source) `