+++ Modules/Microsoft365DSC/DSCResources/MSFT_SCSupervisoryReviewPolicy/MSFT_SCSupervisoryReviewPolicy.psm1
@@ -87,7 +87,7 @@ function Get-TargetResource
             }
             elseif ($null -ne $GlobalAdminAccount)
             {
-                $tenantIdValue = $GlobalAdminAccount.UserName.Split('@')[0]
+                $tenantIdValue = $GlobalAdminAccount.UserName.Split('@')[1]
             }
             Add-M365DSCEvent -Message $_ -EntryType 'Error' `
                 -EventID 1 -Source $($MyInvocation.MyCommand.Source) `
@@ -286,7 +286,7 @@ function Export-TargetResource
             }
             elseif ($null -ne $GlobalAdminAccount)
             {
-                $tenantIdValue = $GlobalAdminAccount.UserName.Split('@')[0]
+                $tenantIdValue = $GlobalAdminAccount.UserName.Split('@')[1]
             }
             Add-M365DSCEvent -Message $_ -EntryType 'Error' `
                 -EventID 1 -Source $($MyInvocation.MyCommand.Source) `