+++ Modules/Microsoft365DSC/DSCResources/MSFT_SPOSiteDesign/MSFT_SPOSiteDesign.psm1
@@ -146,7 +146,7 @@ function Get-TargetResource
             }
             elseif ($null -ne $GlobalAdminAccount)
             {
-                $tenantIdValue = $GlobalAdminAccount.UserName.Split('@')[0]
+                $tenantIdValue = $GlobalAdminAccount.UserName.Split('@')[1]
             }
             Add-M365DSCEvent -Message $_ -EntryType 'Error' `
                 -EventID 1 -Source $($MyInvocation.MyCommand.Source) `
@@ -479,7 +479,7 @@ function Export-TargetResource
             }
             elseif ($null -ne $GlobalAdminAccount)
             {
-                $tenantIdValue = $GlobalAdminAccount.UserName.Split('@')[0]
+                $tenantIdValue = $GlobalAdminAccount.UserName.Split('@')[1]
             }
             Add-M365DSCEvent -Message $_ -EntryType 'Error' `
                 -EventID 1 -Source $($MyInvocation.MyCommand.Source) `