+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Change log for Microsoft365DSC
 
+## 1.20.1118.1
+
+* EXORemoteDomain
+  * Fixed an issue where only non-null parameters are
+    used in the Set-TargetResource resource;
+    (Issue #898)
+* MODULES
+  * M365DSCLogEngine
+    Added Export-M365DiagnosticData function to export diagnostic
+    information to a Zip file.
+
+## 1.20.1111.1
+
+* AADPolicy
+  Initial Release;
+* Fixes an issue with SCRetentionCompliancePolicy where
+  the wrong parameter sets was being passed for creation.
+  (Issue #890)
+
 ## 1.20.1104.1
 
 * AADMSGroup